- CN-106
- CN-106
- Merged PR 6609: CN-106
- rework login page
- Merged PR 6616: P30-959, P30-676 (design changes)
- hor fixes after Ruslan's test:
- Merged PR 6621: hor fixes after Ruslan's test
- P30-989
- rework updatePO mechanism
- Merged PR 6623: P30-989
- statuses sent/save
- Merged PR 6627: statuses sent/save
- move invoiceRecipientAddress into same lofic as invoiceRecipient
- icon fix on login btn
- hot fixes
- Merged PR 6629: P30-992
- hot fix (null -> false)
- Merged PR 6632: hot fix (null -> false)
- font fix
- convert font
- convert font
- change styles and fonts
- login page fixes
- hot fix with filtering (better use filter by boolean instead of .includes + toLower()  for edge cases)
- Merged PR 6647: prod hot fix
- UI fixes on sidenav, anchors, login page
- remove required from fields
- hot fix for prod
- Merged PR 6656: P30-995
- udpate required state of area/type/role without triggering valueChanges
- modal ui fixes
- Merged PR 6667: modal ui fixes
- fix legal contract status log
- P30-999
- link PO to client on creation
- Merged PR 6677: P30-999, P30-1000
- fix edit PO
- Merged PR 6692: fix edit PO
- Merged PR 6693: Merged PR 6692: fix edit PO
- disable/enable address field via 'same as direct client'
- fix data-binding on contracts -> client -> PO
- Merged PR 6704: P30-1007
- Merged PR 6705: Merged PR 6704: P30-1007
- add comment what is left to implement
- emg-widget
- Merged PR 6714: emg-widget
- attachements
- Merged PR 6730: attachements
- stlye fixes after emg-widget merge
- consutlant FA is not selected after save/complete fix (CN-457)
- sticky 'toggle edit mode' and 'return to sales' buttons
- comment out unnecessary css rules (fix errors)
- Merged PR 6731: CN-457, P30-1010
- Merged PR 6732: UI fixes/changes/improvements
- Merge branch 'master' into P30-1003+P30-1004
- add sameAsDirectClient from BE on consultant -> workplace -> onsite
- fix autocomplete behavior
- cleanup
- temporary fix
- Merged PR 6739: P30-1003, P30-1004
- Merged PR 6741: hot workaround to skip legal contracts validation
- Merged PR 6744: Merge branch 'release/2.0.2' into 'master'
- merge fields save
- resolved comments, fixed minor issue
- Merged PR 6751: CN-414 [FE] [Mergefields] Merge fields arbitrarily change from code to result/sample data
- Merge branch 'master' into local-devel
- cleanup after merge

+++ src/app/contracts/shared/editor/editor.component.ts
@@ -1,14 +1,14 @@
 import { ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
 import { CommonModule, Location } from '@angular/common';
 import { ActivatedRoute, Router } from '@angular/router';
-import { catchError, filter, map, mergeMap, pluck, switchMap, take, takeUntil, tap, withLatestFrom } from 'rxjs/operators';
+import { catchError, filter, map, pluck, switchMap, take, takeUntil, tap, withLatestFrom } from 'rxjs/operators';
 import { BehaviorSubject, forkJoin, Observable, of, Subject } from 'rxjs';
 
 // Project Specific
 import { CommentService, CompareService, EditorCoreService } from './services';
 import { RichEditorDirective } from './directives';
 import { RichEditorOptionsProvider } from './providers';
-import { IDocumentItem, IDocumentVersion, IMergeField } from './entities';
+import { IDocumentItem, IDocumentVersion, IMergeField, IMergeFieldState } from './entities';
 
 import { InsertMergeFieldPopupComponent } from './components/insert-merge-field-popup';
 import { CompareSelectVersionPopupComponent } from './components/compare-select-version-popup';
@@ -21,22 +21,18 @@ import { MergeFieldsAbstractService } from './data-access/merge-fields-abstract'
 import { MatButtonModule } from '@angular/material/button';
 import { CommentSidebarComponent } from './components/comment-sidebar/comment-sidebar.component';
 import { inOutPaneAnimation } from './entities/animations';
-import { IntervalApi } from 'devexpress-richedit/lib/model-api/interval';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSelectModule } from '@angular/material/select';
 import { AppCommonModule } from 'src/app/shared/common/app-common.module';
 import { FormControl } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import {
-	AgreementCommentDto,
 	AgreementServiceProxy,
-	AgreementTemplateCommentDto,
 	CompleteTemplateDocumentFileDraftDto,
 	SendDocuSignEnvelopeCommand,
 	SendEmailEnvelopeCommand,
 	StringWrappedValueDto,
 } from 'src/shared/service-proxies/service-proxies';
-import { MatSnackBar } from '@angular/material/snack-bar';
 import { custom } from 'devextreme/ui/dialog';
 import { EditorObserverService } from '../services/editor-observer.service';
 import { SignersPreviewDialogComponent } from 'src/app/workflow/workflow-contracts/legal-contracts/signers-preview-dialog/signers-preview-dialog.component';
@@ -47,6 +43,7 @@ import {
 import { NotificationPopupComponent } from './components/notification-popup';
 import { CommentsAbstractService } from './data-access/comments-abstract.service';
 import { VoidEnvelopePopupComponent } from './components/void-envelope-popup/void-envelope-popup.component';
+import { NotificationType, NotifierService } from './services/notifier.service';
 
 @Component({
 	standalone: true,
@@ -67,12 +64,20 @@ import { VoidEnvelopePopupComponent } from './components/void-envelope-popup/voi
 		MatSelectModule,
 		AppCommonModule,
 	],
-	providers: [RichEditorOptionsProvider, CompareService, CommentService, EditorCoreService, EditorObserverService],
+	providers: [
+		RichEditorOptionsProvider,
+		NotifierService,
+		CompareService,
+		CommentService,
+		EditorCoreService,
+		EditorObserverService,
+	],
 	animations: [inOutPaneAnimation],
 })
 export class EditorComponent implements OnInit, OnDestroy {
 	_destroy$ = new Subject();
 	templateId: number | undefined;
+	notification$ = this._notifierService.notification$;
 	hasUnsavedChanges$ = this._editorCoreService.hasUnsavedChanges$;
 	template$ = new BehaviorSubject<File | Blob | ArrayBuffer | string>(null);
 	documentList$ = new BehaviorSubject<IDocumentItem[]>([]);
@@ -92,6 +97,7 @@ export class EditorComponent implements OnInit, OnDestroy {
 
 	isLoading: boolean = false;
 	isPageLoading: boolean = true;
+	mergeFieldStateBeforeProcessing: IMergeFieldState = null;
 
 	selectedVersionControl = new FormControl();
 
@@ -105,7 +111,7 @@ export class EditorComponent implements OnInit, OnDestroy {
 		private _editorCoreService: EditorCoreService,
 		private _dialog: MatDialog,
 		private _chd: ChangeDetectorRef,
-		private _snackBar: MatSnackBar,
+		private _notifierService: NotifierService,
 		private _editorObserverService: EditorObserverService,
 		private _agreementServiceProxy: AgreementServiceProxy
 	) {}
@@ -132,9 +138,19 @@ export class EditorComponent implements OnInit, OnDestroy {
 			this.mergeFields$.next(res);
 		});
 
-		this.isAgreement$.pipe(take(1)).subscribe(res => {
+		this.isAgreement$.pipe(take(1)).subscribe((res) => {
 			this.isAgreement = !!res;
-		})
+		});
+
+		this.hasUnsavedChanges$.pipe(takeUntil(this._destroy$)).subscribe((hasUnsavedChanges) => {
+			if (hasUnsavedChanges) {
+				this._notifierService.notify(NotificationType.ChangesNotSavedYet);
+			} else {
+				if (this._notifierService.currentNotification?.id === NotificationType.ChangesNotSavedYet) {
+					this._notifierService.notify(NotificationType.Noop);
+				}
+			}
+		});
 	}
 
 	ngOnDestroy(): void {
@@ -332,11 +348,9 @@ export class EditorComponent implements OnInit, OnDestroy {
 	}
 
 	saveCurrentAsDraft() {
-		this.isLoading = true;
-		this._editorCoreService.toggleFields();
-		if (!this.isAgreement) {
-			this._editorCoreService.toggleMergedData();
-		}
+		this.prepareToProcessDocument();
+		let version = this.selectedVersion.version + 1;
+		this._notifierService.notify(NotificationType.VersionBeingCreated, { version });
 
 		this._editorCoreService.setTemplateAsBase64((base64) => {
 			this._agreementService
@@ -345,13 +359,15 @@ export class EditorComponent implements OnInit, OnDestroy {
 					this._updateCommentByNeeds();
 					this.getTemplateVersions(this.templateId);
 					this.cleanUp();
+					this._notifierService.notify(NotificationType.VersionCreatedSuccess, { version });
 				});
 		});
 	}
 
 	saveCurrentAsCompleteAgreementOnly() {
-		this.isLoading = true;
-		this._editorCoreService.toggleFields();
+		this.prepareToProcessDocument();
+		let version = this.selectedVersion.version + 1;
+		this._notifierService.notify(NotificationType.VersionBeingCreated, { version });
 		this._agreementService
 			.unlockAgreementByConfirmation(this.templateId, this.selectedVersion.version)
 			.subscribe((editOpened) => {
@@ -364,20 +380,18 @@ export class EditorComponent implements OnInit, OnDestroy {
 								this._updateCommentByNeeds();
 								this.getTemplateVersions(this.templateId);
 								this.cleanUp();
+								this._notifierService.notify(NotificationType.VersionCreatedSuccess, { version });
 							});
 					});
 				} else {
-					this.isLoading = false;
+					this.cleanUp(true);
+					this._notifierService.notify(NotificationType.ChangesNotSavedYet);
 				}
 			});
 	}
 
 	saveCurrentAsComplete(isAgreement: boolean) {
-		this.isLoading = true;
-		this._editorCoreService.toggleFields();
-		if (!this.isAgreement) {
-			this._editorCoreService.toggleMergedData();
-		}
+		this.prepareToProcessDocument();
 
 		this._editorCoreService.setTemplateAsBase64((base64) => {
 			this._dialog
@@ -399,10 +413,10 @@ export class EditorComponent implements OnInit, OnDestroy {
 				.pipe(
 					map((res) => {
 						if (res) {
+							this._notifierService.notify(NotificationType.SavingChanges);
 							return res;
 						} else {
-							this.isLoading = false;
-							this._chd.detectChanges();
+							this.cleanUp(true);
 							return null;
 						}
 					}),
@@ -420,36 +434,29 @@ export class EditorComponent implements OnInit, OnDestroy {
 				)
 				.subscribe(() => {
 					this.cleanUp();
+					this._notifierService.notify(NotificationType.ChangesSaved);
 				});
 		});
 	}
 
 	saveDraftAsDraft() {
-		this.isLoading = true;
-		this._editorCoreService.toggleFields();
-		if (!this.isAgreement) {
-			this._editorCoreService.toggleMergedData();
-		}
+		this.prepareToProcessDocument();
+		let version = this.currentTemplateVersion;
 
+		this._notifierService.notify(NotificationType.SavingAsADraft, { version });
 		this._editorCoreService.setTemplateAsBase64((base64) => {
 			this._agreementService
 				.saveDraftAsDraftTemplate(this.templateId, false, StringWrappedValueDto.fromJS({ value: base64 }))
 				.subscribe((res) => {
 					this._updateCommentByNeeds();
 					this.cleanUp();
-					if (res) {
-						this.showSnackbar();
-					}
+					this._notifierService.notify(NotificationType.DraftSavedSuccess, { version });
 				});
 		});
 	}
 
 	promoteToDraft() {
-		this.isLoading = true;
-		this._editorCoreService.toggleFields();
-		if (!this.isAgreement) {
-			this._editorCoreService.toggleMergedData();
-		}
+		this.prepareToProcessDocument();
 
 		this._editorCoreService.setTemplateAsBase64((base64) => {
 			this._agreementService
@@ -463,10 +470,12 @@ export class EditorComponent implements OnInit, OnDestroy {
 	}
 
 	saveDraftAsComplete() {
-		this.isLoading = true;
-		this._editorCoreService.toggleFields();
-		if (!this.isAgreement) {
-			this._editorCoreService.toggleMergedData();
+		this.prepareToProcessDocument();
+		let sentVersion = this.versions.find((version) => version.envelopeStatus && version.envelopeStatus === 3);
+		if (sentVersion) {
+			this._notifierService.notify(NotificationType.EnvelopeBeingVoided, { version: sentVersion.version });
+		} else {
+			this._notifierService.notify(NotificationType.SavingChanges);
 		}
 
 		this._editorCoreService.setTemplateAsBase64((base64) => {
@@ -478,16 +487,23 @@ export class EditorComponent implements OnInit, OnDestroy {
 					this.versions
 				)
 				.subscribe((res) => {
-					if (res) {
-						this.showSnackbar();
-						this._updateCommentByNeeds();
-						this.getTemplateVersions(this.templateId);
+					if (!res) {
+						this.cleanUp(true);
+						return this._notifierService.notifyPrevState();
 					}
 
-					this.isLoading = false;
-					this._editorCoreService.removeUnsavedChanges();
 					this.cleanUp();
-					this._chd.detectChanges();
+					this._updateCommentByNeeds();
+					this.getTemplateVersions(this.templateId, () => {
+						this.isLoading = false;
+						if (sentVersion) {
+							this._notifierService.notify(NotificationType.EnvelopeVoidedSuccess, {
+								version: sentVersion.version,
+							});
+						} else {
+							this._notifierService.notify(NotificationType.ChangesSaved);
+						}
+					});
 				});
 		});
 	}
@@ -518,31 +534,30 @@ export class EditorComponent implements OnInit, OnDestroy {
 				});
 
 				dialogRef.componentInstance.onSendViaEmail.subscribe((option: any) => {
-					this._sendViaEmail([this.templateId], true, option);
+					this._sendViaEmail([this.templateId], true, option, result[0].envelopeName);
 				});
 				dialogRef.componentInstance.onSendViaDocuSign.subscribe((option: any) => {
-					this._sendViaDocuSign([this.templateId], true, option);
+					this._sendViaDocuSign([this.templateId], true, option, result[0].envelopeName);
 				});
 			});
 		}
 	}
 
-	private _sendViaEmail(agreementIds: number[], singleEmail: boolean, option: EEmailMenuOption) {
+	private _sendViaEmail(agreementIds: number[], singleEmail: boolean, option: EEmailMenuOption, envelopeName: string) {
+		this._notifierService.notify(NotificationType.SendingInProgress);
 		let input = new SendEmailEnvelopeCommand({
 			agreementIds: agreementIds,
 			singleEmail: singleEmail,
 			convertDocumentFileToPdf: option === EEmailMenuOption.AsPdfFile,
 		});
 		this._agreementServiceProxy.sendEmailEnvelope(input).subscribe(() => {
+			this._notifierService.notify(NotificationType.SentSuccessfully, { filename: envelopeName });
 			this.getTemplateVersions(this.templateId);
-			this._snackBar.open('Successfully sent!', '', {
-				duration: 2500,
-				panelClass: 'green-panel',
-			});
 		});
 	}
 
-	private _sendViaDocuSign(agreementIds: number[], singleEnvelope: boolean, option: EDocuSignMenuOption) {
+	private _sendViaDocuSign(agreementIds: number[], singleEnvelope: boolean, option: EDocuSignMenuOption, envelopeName: string) {
+		this._notifierService.notify(NotificationType.SendingInProgress);
 		let input = new SendDocuSignEnvelopeCommand({
 			agreementIds: agreementIds,
 			singleEnvelope: singleEnvelope,
@@ -550,10 +565,7 @@ export class EditorComponent implements OnInit, OnDestroy {
 		});
 		this._agreementServiceProxy.sendDocusignEnvelope(input).subscribe(() => {
 			this.getTemplateVersions(this.templateId);
-			this._snackBar.open('Successfully sent!', '', {
-				duration: 2500,
-				panelClass: 'green-panel',
-			});
+			this._notifierService.notify(NotificationType.SentSuccessfully, { filename: envelopeName });
 		});
 	}
 
@@ -561,22 +573,46 @@ export class EditorComponent implements OnInit, OnDestroy {
 		this._location.back();
 	}
 
-	cleanUp() {
+	cleanUp(skipUnsavedChangesCheck: boolean = false): void {
 		this.isLoading = false;
 		if (!this.isAgreement) {
 			this._editorCoreService.toggleFields();
 			this._editorCoreService.toggleMergedData();
 		}
+		this.rollbackMergeFieldViewState();
 
-		this._editorCoreService.removeUnsavedChanges();
+		if (!skipUnsavedChangesCheck) {
+			this._editorCoreService.removeUnsavedChanges();
+		}
 		this._chd.detectChanges();
 	}
 
-	showSnackbar() {
-		this._snackBar.open('Successfully saved!', '', {
-			duration: 2500,
-			panelClass: 'green-panel',
-		});
+	prepareToProcessDocument(): void {
+		this.isLoading = true;
+		this._takeMergeFieldStateSnapshot();
+		this._editorCoreService.toggleFields();
+		if (!this.isAgreement) {
+			this._editorCoreService.toggleMergedData();
+		}
+	}
+
+	rollbackMergeFieldViewState(): void {
+		let fieldState = this.mergeFieldStateBeforeProcessing;
+
+		if (fieldState !== undefined || fieldState !== null) {
+			switch (fieldState) {
+				case IMergeFieldState.Field:
+					this._editorCoreService.toggleFields();
+					break;
+				case IMergeFieldState.Code:
+					this._editorCoreService.showAllFieldCodes();
+					break;
+				case IMergeFieldState.Result:
+					this._editorCoreService.showAllFieldResults();
+			}
+
+			this.mergeFieldStateBeforeProcessing = null;
+		}
 	}
 
 	private _showCompleteConfirmDialog(callback: (value: boolean) => void) {
@@ -613,5 +649,9 @@ export class EditorComponent implements OnInit, OnDestroy {
 			.subscribe(() => {
 				this.loadComments(this.templateId);
 			});
+	}
+
+	private _takeMergeFieldStateSnapshot(): void {
+		this.mergeFieldStateBeforeProcessing = this._editorCoreService.mergeFieldState;
 	}
 }
